requirements.txt
pandas
python-dotenv
requests
fastapi
uvicorn
openpyxl
